+++ R/package.R
@@ -57,7 +57,11 @@ env_file <- NULL
 
   env$`__callr_data__`$sofile <- sofile
 
-  env_file <<- tempfile()
+  env_file <<- tempfile("callr-env-")
   saveRDS(env, file = env_file, version = 2, compress = FALSE)
   invisible()
+}
+
+.onUnload <- function(libpath) {
+  if (!is.null(env_file)) unlink(env_file)
 }