+++ R/rscript.R
@@ -30,6 +30,9 @@ rscript <- function(script, cmdargs = character(), libpath = .libPaths(),
 
   options <- setup_rscript_binary_and_args(options)
 
+  ## This cleans up everything...
+  on.exit(unlink(options$tmp_files, recursive = TRUE), add = TRUE)
+
   invisible(run_r(options))
 }
 
@@ -84,7 +87,11 @@ rscript_process <- R6::R6Class(
   inherit = processx::process,
   public = list(
     initialize = function(options)
-      rscript_init(self, private, super, options)
+      rscript_init(self, private, super, options),
+    finalize = function() {
+      unlink(private$options$tmp_files, recursive = TRUE)
+      if ("finalize" %in% ls(super)) super$finalize()
+    }
   ),
   private = list(
     options = NULL