+++ R/utils.R
@@ -106,7 +106,7 @@ bold <- function(x) {
 }
 
 update_history <- function(cmd) {
-  tmp <- tempfile()
+  tmp <- tempfile("callr-hst-")
   on.exit(unlink(tmp, recursive = TRUE))
   utils::savehistory(tmp)
   cat(cmd, "\n", sep = "", file = tmp, append = TRUE)